+++ NEWS.md
@@ -1,4 +1,6 @@
 # nexus 0.1.0.9000
+## New classes and methods
+* Add `split()` methods to divide into groups.
 
 # nexus 0.1.0